Correcao Exercicio 10

+++ IAL-002/2-Seleção/10.py
@@ -2,13 +2,13 @@
 10. Fornecidas duas datas distintas, exibir qual delas ocorre primeiro. Cada data será fornecida em três valores inteiros, onde o primeiro representa o dia, o segundo o mês e o terceiro o ano.
 '''
 print('Data 1')
-d1=float(input("dia: "))
-m1=float(input("mes: "))
-a1=float(input("ano: "))
+d1=int(input("dia: "))
+m1=int(input("mes: "))
+a1=int(input("ano: "))
 print('Data 2')
-d2=float(input("dia: "))
-m2=float(input("mes: "))
-a2=float(input("ano: "))
+d2=int(input("dia: "))
+m2=int(input("mes: "))
+a2=int(input("ano: "))
 if a1>a2:
     print(d2,m2,a2)
 else:
